Updating style
